feat: refactor dashboard page and added loading
